+++ lib/src/date_validators/date_validator_weekday_count_in_month.dart
@@ -1,0 +1,64 @@
+import 'package:equatable/equatable.dart';
+import 'package:time/time.dart';
+
+import '../enums.dart';
+import 'date_validator.dart';
+import 'date_validator_mixin.dart';
+
+/// A [DateValidator] that validates a [DateTime] if the [DateTime.day] is the
+/// [day] of the week and is the [week] of the month.
+class DateValidatorWeekdayCountInMonth extends DateValidator
+    with EquatableMixin, DateValidatorMixin
+    implements Comparable<DateValidatorWeekdayCountInMonth> {
+  /// A [DateValidator] that validates a [DateTime] if the [DateTime.day] is the
+  /// [day] of the week and is the [week] of the month.
+  const DateValidatorWeekdayCountInMonth({
+    required this.week,
+    required this.day,
+  });
+
+  /// A [DateValidator] that validates a [DateTime] if the [DateTime.day] is the
+  /// [day] of the week and is the [week] of the month.
+  factory DateValidatorWeekdayCountInMonth.from(DateTime date) {
+    return DateValidatorWeekdayCountInMonth(
+      week: Week.from(date),
+      day: Weekday.from(date),
+    );
+  }
+
+  /// The expected week of the month.
+  final Week week;
+
+  /// The expected day of the week.
+  final Weekday day;
+
+  @override
+  // ignore: hash_and_equals, already implemented by EquatableMixin
+  bool operator ==(Object other) {
+    return (super == other) ||
+        ((other is DateValidatorWeekdayCountInMonth) &&
+            (week == other.week) &&
+            (day == other.day));
+  }
+
+  @override
+  bool valid(DateTime date) {
+    final valid = week.weekdayOf(
+      year: date.year,
+      month: date.month,
+      day: day,
+      utc: date.isUtc,
+    );
+    return valid.add(date.timeOfDay) == date;
+  }
+
+  @override
+  int compareTo(DateValidatorWeekdayCountInMonth other) {
+    final result = week.compareTo(other.week);
+    if (result == 0) return day.compareTo(other.day);
+    return result;
+  }
+
+  @override
+  List<Object> get props => [day, week];
+}
