+++ lib/src/everies/every_date_validator_intersection.dart
@@ -1,4 +1,4 @@
-import '../date_validator.dart';
+import '../date_validators/date_validators.dart';
 import '../helpers/helpers.dart';
 import 'date_time_limit_reached_exception.dart';
 import 'every_date_validator.dart';
