+++ lib/src/everies/every_due_time_of_day.dart
@@ -1,6 +1,6 @@
 import 'package:time/time.dart';
 
-import '../date_validator.dart';
+import '../date_validators/date_validators.dart';
 import '../extensions.dart';
 import 'every_date_validator.dart';
